refactor agents to separate file
